Fix bugs

+++ app/src/main/java/com/pinkyuni/jogtracker/data/entities/Jog.kt
@@ -2,12 +2,11 @@ package com.pinkyuni.jogtracker.data.entities
 
 import com.google.gson.annotations.SerializedName
 import java.io.Serializable
-import java.math.BigInteger
 
 data class Jog(
-    val id: BigInteger,
+    val id: Long,
     @SerializedName("user_id")
-    val userId: BigInteger,
+    val userId: Long,
     var date: Long,
     var time: Int,
     var distance: Float
